2d examples use common macroquad code
